+++ Senku/Senku/src/Platform/Windows/WindowsInput.h
@@ -9,8 +9,8 @@ namespace Senku
 		virtual bool IsKeyPressedImpl(int keycode) override;
 
 		virtual bool IsMouseButtonPressedImpl(int button) override;
-		virtual float GetMouseXImpl() override;
-		virtual float GetMouseYImpl() override;
-		virtual std::pair<float, float> GetMousePositionImpl() override;
+		virtual double GetMouseXImpl() override;
+		virtual double GetMouseYImpl() override;
+		virtual std::pair<double, double> GetMousePositionImpl() override;
 	};
 }
